backend/requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Banco de dados (MySQL via XAMPP)
sqlalchemy==2.0.30
alembic==1.13.1
aiomysql==0.3.2
pymysql==1.1.2

# Autenticação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Xadrez
python-chess==1.999
stockfish==3.28.0

# HTTP / lichess
httpx==0.27.0

# Dados e visualização
pandas==2.2.2
plotly==5.22.0

# Relatórios PDF
reportlab==4.2.0

# Validação
pydantic==2.7.1
pydantic-settings==2.2.1

# Utilitários
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.4

# Testes
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
